Revert "Merge commit '0ffa3b8' from upstream into servo"
This reverts commit f788437c8fba16c159abe91e315ab2870fc93dc0, reversing
changes made to db27370a1cbafcbfcaeee52a44076a61b3e0573c.

+++ src/events.rs
@@ -1,10 +1,10 @@
-#[derive(Clone, Show, Copy)]
+#[deriving(Clone, Show, Copy)]
 pub enum Event {
     /// The size of the window has changed.
-    Resized(usize, usize),
+    Resized(uint, uint),
 
     /// The position of the window has changed.
-    Moved(isize, isize),
+    Moved(int, int),
 
     /// The window has been closed.
     Closed,
@@ -23,7 +23,7 @@ pub enum Event {
     /// The cursor has moved on the window.
     ///
     /// The parameter are the (x,y) coords in pixels relative to the top-left corner of the window.
-    MouseMoved((isize, isize)),
+    MouseMoved((int, int)),
 
     /// A positive value indicates that the wheel was rotated forward, away from the user;
     ///  a negative value indicates that the wheel was rotated backward, toward the user.
@@ -38,13 +38,13 @@ pub enum Event {
 
 pub type ScanCode = u8;
 
-#[derive(Show, Hash, PartialEq, Eq, Clone, Copy)]
+#[deriving(Show, Hash, PartialEq, Eq, Clone, Copy)]
 pub enum ElementState {
     Pressed,
     Released,
 }
 
-#[derive(Show, Hash, PartialEq, Eq, Clone, Copy)]
+#[deriving(Show, Hash, PartialEq, Eq, Clone, Copy)]
 pub enum MouseButton {
     LeftMouseButton,
     RightMouseButton,
@@ -52,7 +52,7 @@ pub enum MouseButton {
     OtherMouseButton(u8),
 }
 
-#[derive(Show, Hash, PartialEq, Eq, Clone, Copy)]
+#[deriving(Show, Hash, PartialEq, Eq, Clone, Copy)]
 pub enum VirtualKeyCode {
     /// The '1' key over the letters.
     Key1,
